+++ items.lua
@@ -22,33 +22,33 @@ mod.texts = {
 mod.buttons = {
 	--test = {v=true,e=true,x=300,y=20,sx=200,sy=50,func=function(self) print("hay") end}
 	login = {v=true,e=true,x=300,y=400,sx=200,sy=50,func=function(self) login(self.texts.mail.text,self.texts.pass.text) end},
-	exit = {v=true,e=true,x=30,y=100,sx=40,sy=40,func=function(self) if configchanged==true then relogin() configchanged=false end if state == "guestbook" or state == "usermenu" or state == "guestsend" or state == "usermenu" then state = "visit" elseif state == "userlist" then state = "usermenu" else state = "home" substate = "" end end, img="assets/icons/home.png"},
+	exit = {v=true,e=true,x=30,y=100,sx=40,sy=40,func=function(self) if configchanged==true then relogin() configchanged=false end if state == "guestbook" or state == "usermenu" or state == "guestsend" or state == "usermenu" then state = "visit" elseif state == "userlist" then state = "usermenu" elseif state == "topscores" then state="topgame" else state = "home" substate = "" end end, img="assets/icons/home.png"},
 	button = {v=true,e=true,x=515,y=555,sx=40,sy=40,func=function(self) relogin() end, img="assets/icons/refresh.png"}, --refresh
 	button0 = {v=true,e=true,x=0,y=550,sx=100,sy=50,func=function(self) if _G.Client.type == "guest" then state = "login" else logout() end end},
-	button1 = {v=false,e=true,x=110,y=150,sx=100,sy=100,func=function(self) visit() end, img="assets/icons/button1.png"},
-	button2 = {v=false,e=true,x=270,y=150,sx=100,sy=100,func=function(self) if _G.Client.type == "guest" then state = "login" else visit(_G.Client.randomUser()) end end, img="assets/icons/button2.png"},
-	button3 = {v=false,e=true,x=430,y=150,sx=100,sy=100,func=function(self) if _G.Client.type == "guest" then state = "login" else state = "search" end end, img="assets/icons/button3.png"},
-	button4 = {v=false,e=true,x=580,y=150,sx=100,sy=100,func=function(self) substate = "page1" updateTop(_G.Client.topLikes()) end, img="assets/icons/button4.png"},
-	button5 = {v=false,e=true,x=580,y=310,sx=100,sy=100,func=function(self) state = "conf" end, img="assets/icons/button5.png"},
-	--button6 = {v=false,e=true,x=110,y=310,sx=100,sy=100,func=function(self) state = "topgame" end, img="assets/icons/button6.png"},
-	button7 = {v=false,e=true,x=270,y=310,sx=100,sy=100,func=function(self) if _G.Client.type == "guest" then state = "login" else state = "tictaclobby" substate="page1" updateGame(_G.Client.getGameSessions(drawPou.toTable(_G.Client.me).i,8,1,100)) end end, img="assets/icons/button7.png"},
-	button8 = {v=false,e=true,x=430,y=310,sx=100,sy=100,func=function(self) if _G.Client.type == "guest" then state = "login" else state = "tictaclobby" substate="page1" updateGame(_G.Client.getGameSessions(drawPou.toTable(_G.Client.me).i,9,1,100)) end end, img="assets/icons/button8.png"},
-	zakehweb = {v=false,e=true,x=750,y=550,sx=50,sy=50,func=function(self) love.system.openURL(server.links.web) end, img="assets/icons/zakeh.png"},
-	instagram = {v=false,e=true,x=690,y=550,sx=50,sy=50,func=function(self) love.system.openURL("https://www.instagram.com/"..server.links.instagram) end, img="assets/icons/insta.png"},
-	facebook = {v=false,e=true,x=630,y=550,sx=50,sy=50,func=function(self) love.system.openURL("https://www.facebook.com/"..server.links.facebook) end, img="assets/icons/fb.png"},
-	twitter = {v=false,e=true,x=570,y=550,sx=50,sy=50,func=function(self) love.system.openURL("https://www.twitter.com/"..server.links.twitter) end, img="assets/icons/twitter.png"},
+	button1 = {v=false,e=true,x=110,y=150,sx=100,sy=100,func=function(self) visit() end, img="assets/icons/button1.png",hint="Bmypou"},
+	button2 = {v=false,e=true,x=270,y=150,sx=100,sy=100,func=function(self) if _G.Client.type == "guest" then state = "login" else visit(_G.Client.randomUser()) end end, img="assets/icons/button2.png",hint="Brandompou"},
+	button3 = {v=false,e=true,x=430,y=150,sx=100,sy=100,func=function(self) if _G.Client.type == "guest" then state = "login" else state = "search" end end, img="assets/icons/button3.png",hint="Bsearch"},
+	button4 = {v=false,e=true,x=580,y=150,sx=100,sy=100,func=function(self) substate = "page1" updateTop(_G.Client.topLikes()) end, img="assets/icons/button4.png",hint="Btoplikes"},
+	button5 = {v=false,e=true,x=580,y=310,sx=100,sy=100,func=function(self) state = "conf" end, img="assets/icons/button5.png",hint="Bsettings"},
+	button6 = {v=false,e=true,x=110,y=310,sx=100,sy=100,func=function(self) state = "topgame" end, img="assets/icons/button6.png",hint="Btopgames"},
+	button7 = {v=false,e=true,x=270,y=310,sx=100,sy=100,func=function(self) if _G.Client.type == "guest" then state = "login" else state = "tictaclobby" substate="page1" updateGame(_G.Client.getGameSessions(drawPou.toTable(_G.Client.me).i,8,1,100)) end end, img="assets/icons/button7.png",hint="Btictac"},
+	button8 = {v=false,e=true,x=430,y=310,sx=100,sy=100,func=function(self) if _G.Client.type == "guest" then state = "login" else state = "tictaclobby" substate="page1" updateGame(_G.Client.getGameSessions(drawPou.toTable(_G.Client.me).i,9,1,100)) end end, img="assets/icons/button8.png",hint="Bfourpous"},
+	zakehweb = {v=false,e=true,x=750,y=550,sx=50,sy=50,func=function(self) love.system.openURL(server.links.web) end, img="assets/icons/zakeh.png",hint="Bweb"},
+	instagram = {v=false,e=true,x=690,y=550,sx=50,sy=50,func=function(self) love.system.openURL("https://www.instagram.com/"..server.links.instagram) end, img="assets/icons/insta.png",hint="Binsta"},
+	facebook = {v=false,e=true,x=630,y=550,sx=50,sy=50,func=function(self) love.system.openURL("https://www.facebook.com/"..server.links.facebook) end, img="assets/icons/fb.png",hint="Bfaceb"},
+	twitter = {v=false,e=true,x=570,y=550,sx=50,sy=50,func=function(self) love.system.openURL("https://www.twitter.com/"..server.links.twitter) end, img="assets/icons/twitter.png",hint="Btwit"},
 	like = {v=true,e=true,x=80,y=320,sx=100,sy=100,func=function(self) if account.isFollowing==1 then unlike() else like() end end},
 	usermenu = {v=true,e=true,x=575,y=262.5,sx=50,sy=50,func=function(self) state = "usermenu" end, img="assets/icons/menu.png"},
-	likes = {v=true,e=true,x=270,y=150,sx=100,sy=100,func=function(self) updateTop(_G.Client.getLikers(account.id,0),"userlist") end, img="assets/icons/likes_button.png"},
-	follows = {v=true,e=true,x=110,y=150,sx=100,sy=100,func=function(self) updateTop(_G.Client.getFavorites(account.id,0),"userlist") end, img="assets/icons/follows_button.png"},
-	visitors = {v=true,e=true,x=430,y=150,sx=100,sy=100,func=function(self) updateTop(_G.Client.getVisitors(account.id,0),"userlist") end, img="assets/icons/visits_button.png"},
+	likes = {v=true,e=true,x=270,y=150,sx=100,sy=100,func=function(self) updateTop(_G.Client.getLikers(account.id,0),"userlist") end, img="assets/icons/likes_button.png",hint="Bfollowers"},
+	follows = {v=true,e=true,x=110,y=150,sx=100,sy=100,func=function(self) updateTop(_G.Client.getFavorites(account.id,0),"userlist") end, img="assets/icons/follows_button.png",hint="Bfollowing"},
+	visitors = {v=true,e=true,x=430,y=150,sx=100,sy=100,func=function(self) updateTop(_G.Client.getVisitors(account.id,0),"userlist") end, img="assets/icons/visits_button.png",hint="Bvisits"},
 	guestbook = {v=true,e=true,x=650,y=262.5,sx=50,sy=50,func=function(self) getMsgs(account.id) end, img="assets/icons/guestbook.png"},
 	gbmenu = {v=false,e=true,x=90,y=83,sx=620,sy=42,func=function(self) state = "guestsend" substate="page1" end},
-	sNick = {v=true,e=true,x=330,y=170,sx=50,sy=50,func=function(self) visit(_G.Client.getUserByNickname(self.texts.sNick.text)) end},
-	miniNick = {v=true,e=true,x=380,y=170,sx=50,sy=50,func=function(self) visit(_G.Client.getAvatarByNickname(self.texts.sNick.text),true) end,hint=translate:Get("searchmini")},
-	sMail = {v=true,e=true,x=330,y=270,sx=50,sy=50,func=function(self) visit(_G.Client.getUserByEmail(self.texts.sMail.text)) end},
-	miniMail = {v=true,e=true,x=380,y=270,sx=50,sy=50,func=function(self) visit(_G.Client.getAvatarByEmail(self.texts.sMail.text),true) end,hint=translate:Get("searchmini")},
-	sID = {v=true,e=true,x=330,y=370,sx=50,sy=50,func=function(self) visit(_G.Client.getUserById(self.texts.sID.text)) end},
+	sNick = {v=true,e=true,x=330,y=170,sx=50,sy=50,func=function(self) visit(_G.Client.getUserByNickname(self.texts.sNick.text)) end,img="assets/icons/search.png"},
+	miniNick = {v=true,e=true,x=380,y=170,sx=50,sy=50,func=function(self) visit(_G.Client.getAvatarByNickname(self.texts.sNick.text),true) end,hint="searchmini",img="assets/icons/search_fast.png"},
+	sMail = {v=true,e=true,x=330,y=270,sx=50,sy=50,func=function(self) visit(_G.Client.getUserByEmail(self.texts.sMail.text)) end,img="assets/icons/search.png"},
+	miniMail = {v=true,e=true,x=380,y=270,sx=50,sy=50,func=function(self) visit(_G.Client.getAvatarByEmail(self.texts.sMail.text),true) end,hint="searchmini",img="assets/icons/search_fast.png"},
+	sID = {v=true,e=true,x=330,y=370,sx=50,sy=50,func=function(self) visit(_G.Client.getUserById(self.texts.sID.text)) end,img="assets/icons/search.png"},
 	right = {v=true,e=true,x=730,y=290,sx=40,sy=40,func=function(self) local _,__ = string.gsub(substate,"page","") substate="page".. (tonumber(_) or 1)+1 end, img="assets/icons/next.png"},
 	left = {v=true,e=true,x=30,y=290,sx=40,sy=40,func=function(self) local _,__ = string.gsub(substate,"page","") substate="page".. (tonumber(_) or 1)-1 end, img="assets/icons/prev.png"},
 	conf1 = {v=true,e=true,x=650,y=370,sx=30,sy=30,func=function(self) changeHost(self.texts.host.text) end}, --change host
@@ -58,7 +58,11 @@ mod.buttons = {
 	conf6 = {v=true,e=true,x=330,y=490,sx=30,sy=30,func=function(self) translate:Next() end}, --next lang
 	conf7 = {v=true,e=true,x=130,y=490,sx=30,sy=30,func=function(self) translate:Prev() end}, --prev lang
 	about = {v=false,e=true,x=450,y=475,sx=180,sy=60,func=function(self) state = "about" end, img="assets/icons/about.png"}, --about
-	git = {v=true,e=true,x=610,y=150,sx=50,sy=50,func=function(self) love.system.openURL("https://www.github.com/h401217/pou-search") end, img="assets/icons/github.png"}, --git
+	git = {v=true,e=true,x=610,y=150,sx=50,sy=50,func=function(self) love.system.openURL("https://www.github.com/h401217/pou-search") end, img="assets/icons/github.png",hint="Bgit"}, --git
+	gday1 = {v=false,e=true,x=124,y=90,sx=138,sy=30,func=function(self) extra.options.gtopday = "today" end},
+	gday2 = {v=false,e=true,x=262,y=90,sx=138,sy=30,func=function(self) extra.options.gtopday = "week" end},
+	gday3 = {v=false,e=true,x=400,y=90,sx=138,sy=30,func=function(self) extra.options.gtopday = "month" end},
+	gday4 = {v=false,e=true,x=538,y=90,sx=138,sy=30,func=function(self) extra.options.gtopday = "alltime" end},
 }
 
 --
@@ -66,6 +70,12 @@ mod.buttons = {
 		for d=0,1 do
 			local position = ((d+1)+(2*c))
 			mod.buttons["pou_user"..position] = {v=false,e=true,x=90+330*d,y=130+90*c,sx=290,sy=80,func=function(self) clickuser(position) end}
+		end
+	end
+	for d=0,4 do
+		for c=0,6 do
+			local position = (c+1)+(7*d)
+			mod.buttons["pou_game"..position] = {v=false,e=true,x=85+90*c,y=130+90*d,sx=90,sy=90,func=function(self) clickgame(position) end, hint=extra.games[position]}
 		end
 	end
 --
@@ -136,6 +146,7 @@ function mod:draw()
 	local mX,mY = love.mouse.getPosition()
 	local oldfont = love.graphics.getFont()
 	local cFont = love.graphics.newFont(10)
+	local dFont = love.graphics.newFont(12)
 	for a,b in pairs(mod.texts) do
 		if b.v == true then
 			love.graphics.setFont(cFont)
@@ -178,14 +189,17 @@ function mod:draw()
 		if b.e == true then
 			if b.hint then
 				if mX >= b.x and mX <= b.x+b.sx and mY >= b.y and mY <= b.y+b.sy and b.e == true then
-					love.graphics.setFont(cFont)
-					local lines = (select(2,string.gsub(b.hint,"\n","")) or 0)+1
+					love.graphics.setFont(dFont)
+					local txt = translate:Get(b.hint)
+					local lines = (select(2,string.gsub(txt,"\n","")) or 0)+1
 					local off = 5
-					local w = cFont:getWidth(b.hint)+(off*2)
-					local h = cFont:getHeight(b.hint)+(off*2)
-					love.graphics.setColor(0.92,0.85,0.3,1) love.graphics.rectangle("fill",mX,mY-h,w,h)
-					love.graphics.setColor(0.9,0.6,0.07,1) love.graphics.rectangle("line",mX,mY-h,w,h)
-					love.graphics.setColor(0,0,0,1) love.graphics.print(b.hint,(mX+(w/2))-cFont:getWidth(b.hint)/2,(mY-(h/2))-cFont:getHeight(b.hint)/2)
+					local w = dFont:getWidth(txt)+(off*2)
+					local h = dFont:getHeight(txt)+(off*2)
+					local offx = (mX+w > 800) and (mX+w-800) or 0
+					
+					love.graphics.setColor(0.92,0.85,0.3,1) love.graphics.rectangle("fill",mX-offx,mY-h,w,h)
+					love.graphics.setColor(0.9,0.6,0.07,1) love.graphics.rectangle("line",mX-offx,mY-h,w,h)
+					love.graphics.setColor(0,0,0,1) love.graphics.print(txt,(mX+((w/2))-dFont:getWidth(txt)/2)-offx,(mY-(h/2))-dFont:getHeight(txt)/2)
 					love.graphics.setColor(1,1,1,1)
 				end
 			end